requirements.txt
selene >= 2.0.0rc7
pytest
allure-pytest==2.13.2